Move folding and visiting traits into type library
